+++ frontend/src/components/EntityEdit.vue
@@ -48,7 +48,8 @@ Modal windows to create, delete and edit entities
                     <p>Are you sure you want to <strong>permanently</strong> delete this {{ type }}?</p>
                     <p class="text-warning">{{ item.name ? item.name : item.id }}</p>
                     <p v-cloak v-if="type === 'EveLogin'">
-                        !!! This will also delete <strong>all</strong> associated existing tokens from any character.
+                        <span role="img" class="fas fa-exclamation-triangle text-warning"></span>
+                        This will also delete <strong>all</strong> associated existing tokens from any character!
                     </p>
                 </div>
                 <div class="modal-footer">